Initial travis setup
Setting up travis-CI with stage builds (BETA). Sudo is needed to install
all the tools and make them run-able in Travis. Since Docker usage is
harder in Travis-CI that meant quite some fiddling around and on the end
custom LCOV needed to be downloaded and installed for coverage to work.
In near future we should add more jobs for both tasks.

Adding Travis badge to README.